re-arrange code
* create mdms-metanome module for Metanome integration
* remove Location implementations from mdms-flink
* refactor the Flink adapters (cell and tuple data source builders)
