--- mdms/mdms-flink/src/main/java/de/hpi/isg/mdms/flink/location/FileLocation.java
@@ -1,34 +1,0 @@
-/***********************************************************************************************************************
- * Copyright (C) 2014 by Sebastian Kruse
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- **********************************************************************************************************************/
-package de.hpi.isg.mdms.flink.location;
-
-import de.hpi.isg.mdms.model.location.DefaultLocation;
-import org.apache.flink.core.fs.Path;
-
-/**
- *
- *
- * @author Sebastian Kruse
- */
-@SuppressWarnings("serial")
-public class FileLocation extends DefaultLocation {
-    
-	public void setPath(String path) {
-		set(PATH, path);
-	}
-	
-    public Path getPath() {
-        return new Path(get(PATH));
-    }
-    
-}
