+++ mdms/mdms-metanome/src/main/java/de/hpi/isg/mdms/metanome/json/FunctionalDependencyReader.java
@@ -1,0 +1,21 @@
+package de.hpi.isg.mdms.metanome.json;
+
+import de.hpi.isg.mdms.metanome.DependencyResultReceiver;
+import de.metanome.algorithm_integration.result_receiver.CouldNotReceiveResultException;
+import de.metanome.algorithm_integration.results.FunctionalDependency;
+
+/**
+ * Parser for {@link FunctionalDependency} constraints.
+ */
+public class FunctionalDependencyReader extends AbstractJsonReader<FunctionalDependency> {
+
+    @Override
+    protected void feed(FunctionalDependency metanomeConstraint, DependencyResultReceiver<?> resultReceiver) throws CouldNotReceiveResultException {
+        resultReceiver.receiveResult(metanomeConstraint);
+    }
+
+    @Override
+    protected Class<FunctionalDependency> getMetanomeTypeClass() {
+        return FunctionalDependency.class;
+    }
+}
