+++ mdms/mdms-tools/pom.xml
@@ -26,19 +26,6 @@
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
         </dependency>
-
-        <dependency>
-            <groupId>de.metanome</groupId>
-            <artifactId>algorithm_integration</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>de.metanome</groupId>
-            <artifactId>algorithm_helper</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>de.metanome</groupId>
-            <artifactId>backend</artifactId>
-        </dependency>
         <dependency>
             <groupId>org.antlr</groupId>
             <artifactId>antlr4-runtime</artifactId>
@@ -57,6 +44,11 @@
         <dependency>
             <groupId>de.hpi.isg</groupId>
             <artifactId>mdms-util</artifactId>
+            <version>0.0.3-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>de.hpi.isg</groupId>
+            <artifactId>mdms-metanome</artifactId>
             <version>0.0.3-SNAPSHOT</version>
         </dependency>
         <dependency>
