+++ mdms/mdms-tools/src/main/java/de/hpi/isg/mdms/tools/apps/GenerateMetadataStoreApp.java
@@ -16,11 +16,11 @@ import com.beust.jcommander.Parameter;
 import com.beust.jcommander.ParametersDelegate;
 import de.hpi.isg.mdms.Encoding;
 import de.hpi.isg.mdms.clients.apps.AppTemplate;
+import de.hpi.isg.mdms.clients.location.CsvFileLocation;
 import de.hpi.isg.mdms.clients.parameters.JCommanderParser;
 import de.hpi.isg.mdms.clients.parameters.MetadataStoreParameters;
 import de.hpi.isg.mdms.clients.util.MetadataStoreUtil;
 import de.hpi.isg.mdms.domain.constraints.*;
-import de.hpi.isg.mdms.flink.location.CsvFileLocation;
 import de.hpi.isg.mdms.model.MetadataStore;
 import de.hpi.isg.mdms.model.constraints.ConstraintCollection;
 import de.hpi.isg.mdms.model.targets.Column;
