--- mdms/mdms-tools/src/main/java/de/hpi/isg/mdms/tools/metanome/friendly/UniqueColumnCombinationReader.java
@@ -1,28 +1,0 @@
-package de.hpi.isg.mdms.tools.metanome.friendly;
-
-import de.hpi.isg.mdms.tools.metanome.DependencyResultReceiver;
-import de.metanome.algorithm_integration.result_receiver.CouldNotReceiveResultException;
-import de.metanome.algorithm_integration.results.UniqueColumnCombination;
-import de.metanome.backend.result_receiver.ResultReceiver;
-
-import java.util.regex.Pattern;
-
-/**
- * Parser for {@link UniqueColumnCombination} constraints.
- */
-public class UniqueColumnCombinationReader extends AbstractFriendlyReader<UniqueColumnCombination> {
-
-    @Override
-    protected void processLine(String line, DependencyResultReceiver<?> resultReceiver) {
-        try {
-            resultReceiver.receiveResult(toUCC(line));
-        } catch (CouldNotReceiveResultException e) {
-            throw new RuntimeException(String.format("Could not process \"{}\".", line), e);
-        }
-    }
-
-    private static UniqueColumnCombination toUCC(String line) {
-        String cleanLine = line.replaceAll(Pattern.quote("[") + "|" + Pattern.quote("]"), "");
-        return new UniqueColumnCombination(toColumnCombination(cleanLine));
-    }
-}
